+++ src/features/investorRelations/pages/CodeConductPage/CodeConductPage.tsx
@@ -11,6 +11,7 @@ import { AnimatePresence } from 'framer-motion'
 import { Animation } from '@/components/Animation'
 import { useRouter } from '@/libs/intl/navigation'
 import GroupAccordionTabs from '@/features/investorRelations/components/GroupAccordionTabs/GroupAccordionTabs'
+import { ContactUsForm } from '@/components/ContactUs/ContactUsForm'
 
 export function CodeOfConductPage({
   conOdConductData,
@@ -35,11 +36,14 @@ export function CodeOfConductPage({
             alt='investor-banner-4'
           />
 
-          <section className='px-5 pb-5  pt-[50px] md:pt-[100px] max-w-4xl mx-auto space-y-6'>
+          <section className='px-5 pb-5 pt-[50px] md:pt-[100px] max-w-4xl mx-auto space-y-6'>
             <h1 className='text-lg md:text-3xl mb-10 text-blue-400 text-center'>
               {t('title')}
             </h1>
             <GroupAccordionTabs pageKey='code-of-conduct' group={group} />
+          </section>
+          <section className='px-5 pb-5 pt-[50px] max-w-4xl mx-auto space-y-6'>
+            <ContactUsForm hideContactForm />
           </section>
         </Animation>
       </AnimatePresence>
